feat(LESQ-1002): share layouts between onboarding and registration
makes it easier to line things up
